README.md: update __init__.py: add this to make it a module config.ini: remove gid as this is found in utils.py daemon.py: * split functionality into new files `qstat.py` and `utils.py` * command-line arg for config file
