+++ config.ini
@@ -1,7 +1,6 @@
 [Users]
-list = pde3,jll1062,gut10,auk37,dup193,dfc13,fxo10,fxh140,tba109,mxw5234,sxw530,kjc5697
+list = auk37,dfc13,dup193,fxh140,fxo10,gut10,jll1062,kjc5697,mxw5234,pde3,sxw530,tba109
 group = dfc13_collab
-gid = 89100
 
 [Queue]
 n_run = 260
